# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundBoard - Community Music PWA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SongSource = typeof SongSource[keyof typeof SongSource];


export const SongSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
  upload: 'upload',
} as const;

export type SongStorageType = typeof SongStorageType[keyof typeof SongStorageType];


export const SongStorageType = {
  limited: 'limited',
  permanent: 'permanent',
  public_limited: 'public_limited',
  public_download: 'public_download',
} as const;

export interface Song {
  id: number;
  title: string;
  artist: string;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  coverUrl?: string | null;
  source: SongSource;
  sourceUrl: string;
  storageType: SongStorageType;
  category: string;
  tags?: string[];
  userId: string;
  /** @nullable */
  expiresAt?: string | null;
  voteCount?: number;
  isPublic?: boolean;
  createdAt: string;
}

export type SongInputSource = typeof SongInputSource[keyof typeof SongInputSource];


export const SongInputSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
  upload: 'upload',
} as const;

export type SongInputStorageType = typeof SongInputStorageType[keyof typeof SongInputStorageType];


export const SongInputStorageType = {
  limited: 'limited',
  permanent: 'permanent',
  public_limited: 'public_limited',
  public_download: 'public_download',
} as const;

export interface SongInput {
  title: string;
  artist: string;
  album?: string;
  duration?: number;
  coverUrl?: string;
  source: SongInputSource;
  sourceUrl: string;
  storageType: SongInputStorageType;
  category?: string;
  tags?: string[];
  userId: string;
  isPublic?: boolean;
}

export type SongUpdateStorageType = typeof SongUpdateStorageType[keyof typeof SongUpdateStorageType];


export const SongUpdateStorageType = {
  limited: 'limited',
  permanent: 'permanent',
  public_limited: 'public_limited',
  public_download: 'public_download',
} as const;

export interface SongUpdate {
  title?: string;
  artist?: string;
  album?: string;
  coverUrl?: string;
  category?: string;
  tags?: string[];
  storageType?: SongUpdateStorageType;
}

export type PromoteInputTargetStorage = typeof PromoteInputTargetStorage[keyof typeof PromoteInputTargetStorage];


export const PromoteInputTargetStorage = {
  permanent: 'permanent',
  public_download: 'public_download',
} as const;

export interface PromoteInput {
  targetStorage: PromoteInputTargetStorage;
}

export interface TagInput {
  tag: string;
  userId: string;
}

export type SearchResultSource = typeof SearchResultSource[keyof typeof SearchResultSource];


export const SearchResultSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
} as const;

export interface SearchResult {
  externalId: string;
  title: string;
  artist: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  coverUrl?: string | null;
  source: SearchResultSource;
  streamUrl: string;
}

export type ImportInputSource = typeof ImportInputSource[keyof typeof ImportInputSource];


export const ImportInputSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
} as const;

export interface ImportInput {
  externalId: string;
  title: string;
  artist: string;
  duration?: number;
  coverUrl?: string;
  source: ImportInputSource;
  streamUrl: string;
  userId: string;
  category?: string;
}

export interface Category {
  name: string;
  count: number;
}

export interface Playlist {
  id: number;
  name: string;
  userId: string;
  isPublic: boolean;
  songCount: number;
  songs?: Song[];
  createdAt: string;
}

export interface PlaylistInput {
  name: string;
  userId: string;
  isPublic?: boolean;
}

export interface PlaylistSongInput {
  songId: number;
}

export interface BoardEntry {
  id: number;
  songId: number;
  song: Song;
  voteCount: number;
  nominatedBy: string;
  userVoted?: boolean;
  createdAt: string;
}

export interface VoteInput {
  userId: string;
}

export interface NominationInput {
  userId: string;
}

export interface QueueEntry {
  id: number;
  songId: number;
  song: Song;
  userId: string;
  vetoCount: number;
  position: number;
  createdAt: string;
}

export interface QueueInput {
  songId: number;
  userId: string;
}

export interface VetoInput {
  userId: string;
}

export interface DailyEntry {
  id: number;
  songId: number;
  song: Song;
  userId: string;
  date: string;
  createdAt: string;
}

export interface DailyInput {
  songId: number;
  userId: string;
}

export interface ForumMessage {
  id: number;
  userId: string;
  /** @nullable */
  userName?: string | null;
  content: string;
  /** @nullable */
  moodTag: string | null;
  createdAt: string;
}

export interface ForumMessageInput {
  userId: string;
  content: string;
  moodTag?: string;
}

export interface ForumPage {
  messages: ForumMessage[];
  total: number;
  page: number;
  totalPages: number;
}

export interface UserProfile {
  userId: string;
  displayName: string;
  reputation: number;
  badgeCount: number;
  badges?: string[];
  nominatedCount?: number;
  uploadCount?: number;
  createdAt: string;
}

export interface UserInput {
  displayName: string;
}

export interface InboxItem {
  id: number;
  fromUserId: string;
  /** @nullable */
  fromUserName?: string | null;
  toUserId: string;
  songId: number;
  song: Song;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface ShareInput {
  fromUserId: string;
  toUserId: string;
  songId: number;
  message?: string;
}

export interface AppStats {
  totalSongs: number;
  totalUsers: number;
  totalVotes: number;
  activeDailyEntries: number;
  queueLength: number;
  forumMessages: number;
  publicDownloadCount?: number;
  publicLimitedCount?: number;
}

export type ListSongsParams = {
category?: string;
storageType?: string;
sort?: string;
search?: string;
limit?: number;
offset?: number;
};

export type SearchYoutubeParams = {
q: string;
};

export type SearchSoundcloudParams = {
q: string;
};

export type ListPlaylistsParams = {
userId?: string;
};

export type ListSongBoardParams = {
tab?: ListSongBoardTab;
limit?: number;
offset?: number;
};

export type ListSongBoardTab = typeof ListSongBoardTab[keyof typeof ListSongBoardTab];


export const ListSongBoardTab = {
  hot: 'hot',
  legends: 'legends',
  mini: 'mini',
} as const;

export type GetDailyArchiveParams = {
date?: string;
limit?: number;
};

export type ListForumMessagesParams = {
moodTag?: string;
page?: number;
limit?: number;
};

export type GetTrendingParams = {
limit?: number;
};