# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SoundBoard - Community Music PWA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List user playlists
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get playlist with songs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists/{id}/songs:
    post:
      operationId: addSongToPlaylist
      tags: [playlists]
      summary: Add a song to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistSongInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}/songs/{songId}:
    delete:
      operationId: removeSongFromPlaylist
      tags: [playlists]
      summary: Remove a song from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Song Board ─────────────────────────────────────────────────────────────
...
  /queue:
    get:
      operationId: getQueue
      tags: [queue]
      summary: Get the current mixed queue
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add a song to the mixed queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{id}/veto:
    post:
      operationId: vetoQueueEntry
      tags: [queue]
      summary: Veto/skip a song in the mixed queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VetoInput"
      responses:
        "200":
          description: Vetoed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{id}:
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove a song from the queue (after playing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Daily Playlist ─────────────────────────────────────────────────────────
...
  /daily/submit:
    post:
      operationId: submitToDaily
      tags: [daily]
      summary: Submit a song to today's daily playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyEntry"
  /daily/archive:
    get:
      operationId: getDailyArchive
      tags: [daily]
      summary: Get archived daily playlists
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Archived dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyEntry"

  # ── Recommendation Board (Forum) ───────────────────────────────────────────
  /forum:
    get:
      operationId: listForumMessages
      tags: [forum]
      summary: List forum messages (paginated)
      parameters:
        - name: moodTag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Forum messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPage"
    post:
      operationId: createForumMessage
      tags: [forum]
      summary: Post a message to the recommendation board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumMessage"

  # ── User Profiles ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artist, source, sourceUrl, storageType, category, userId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        coverUrl:
          type: ["string", "null"]
        source:
          type: string
          enum: [youtube, soundcloud, upload]
        sourceUrl:
          type: string
        storageType:
          type: string
          enum: [limited, permanent, public_limited, public_download]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        userId:
          type: string
        expiresAt:
          type: ["string", "null"]
        voteCount:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
...
    Playlist:
      type: object
      required: [id, name, userId, isPublic, songCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        userId:
          type: string
        isPublic:
          type: boolean
        songCount:
          type: integer
        songs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        createdAt:
          type: string
    PlaylistInput:
      type: object
      required: [name, userId]
      properties:
        name:
          type: string
        userId:
          type: string
        isPublic:
          type: boolean
    PlaylistSongInput:
      type: object
      required: [songId]
      properties:
        songId:
          type: integer

    # ── Song Board ─────────────────────────────────────────────────────────
...
    QueueEntry:
      type: object
      required: [id, songId, song, userId, vetoCount, position, createdAt]
      properties:
        id:
          type: integer
        songId:
          type: integer
        song:
          $ref: "#/components/schemas/Song"
        userId:
          type: string
        vetoCount:
          type: integer
        position:
          type: integer
        createdAt:
          type: string
    QueueInput:
      type: object
      required: [songId, userId]
      properties:
        songId:
          type: integer
        userId:
          type: string
    VetoInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: string

    # ── Daily Playlist ─────────────────────────────────────────────────────
    DailyEntry:
      type: object
      required: [id, songId, song, userId, date, createdAt]
      properties:
        id:
          type: integer
        songId:
          type: integer
        song:
          $ref: "#/components/schemas/Song"
        userId:
          type: string
        date:
          type: string
        createdAt:
          type: string
    DailyInput:
      type: object
      required: [songId, userId]
      properties:
        songId:
          type: integer
        userId:
          type: string

    # ── Forum ──────────────────────────────────────────────────────────────
    ForumMessage:
      type: object
      required: [id, userId, content, moodTag, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        userName:
          type: ["string", "null"]
        content:
          type: string
        moodTag:
          type: ["string", "null"]
        createdAt:
          type: string
    ForumMessageInput:
      type: object
      required: [userId, content]
      properties:
        userId:
          type: string
        content:
          type: string
        moodTag:
          type: string
    ForumPage:
      type: object
      required: [messages, total, page, totalPages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ForumMessage"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer

    # ── Users ──────────────────────────────────────────────────────────────
...